Move MediaProgressCoordinator class to own file.
